Name Change
Changed the name of main.css to styles.css and updated all links in html documents. github pages still not showing my website with css. however if i modify the code on the live website with the right path it works. unsure why it continues to fail building and deploying the website. Error 400 is the only code shown